Move checkIfInt() from historynode module to helper module
