+++ src/connection.py
@@ -1,6 +1,6 @@
 """ Stores connection info between game nodes (positions) """
 
-from src import historynode
+from src import helper
 
 class Connection(object):
     """ Stores connection info between game nodes (positions) """
@@ -14,25 +14,25 @@ class Connection(object):
 
     def setstartX(self, value):
         """ Setter for startX with type checking """
-        historynode.checkIfInt(value)
+        helper.checkIfInt(value)
         self.startX = value
 
     def setstartY(self, value):
         """ Setter for startY with type checking """
-        historynode.checkIfInt(value)
+        helper.checkIfInt(value)
         self.startY = value
 
     def setdirection(self, value):
         """ Setter for direction with type checking """
-        historynode.checkIfInt(value)
+        helper.checkIfInt(value)
         self.direction = value
 
     def setendX(self, value):
         """ Setter for endX with type checking """
-        historynode.checkIfInt(value)
+        helper.checkIfInt(value)
         self.endX = value
 
     def setendY(self, value):
         """ Setter for endY with type checking """
-        historynode.checkIfInt(value)
+        helper.checkIfInt(value)
         self.endY = value
