+++ test/test_historynode.py
@@ -174,14 +174,6 @@ class TestHistoryNode(unittest.TestCase):
         actual_result = hn_obj.foxesWinP()
         self.assertEqual(actual_result, expected_result)
 
-    def test_checkIfInt_good(self):
-        """ Check that checkIfInt correctly assesses an input """
-        self.assertFalse(historynode.checkIfInt(123))
-
-    def test_checkIfInt_bad(self):
-        """ Check that checkIfInt raises an error with non-int input """
-        self.assertRaises(TypeError, historynode.checkIfInt, "abc")
-
     def test_setGameType_good(self):
         """ Check that gameType is set correctly with valid input """
         hn_obj = historynode.HistoryNode()
